[svn-r6131] Purpose:     Code cleanup
Description:
    Clean up compiler warnings from IRIX64 builds.

Platforms tested:
    IRIX64 6.5 (modi4)

+++ src/H5I.c
@@ -1168,7 +1168,7 @@ H5Iget_name(hid_t id, char *name/*out*/, size_t size)
     } /* end if */
 
     /* Set return value */
-    ret_value=len;
+    ret_value=(ssize_t)len;
 
 done:
     FUNC_LEAVE(ret_value);
